created a start class and moved all the functions that were in Bank to it

+++ src/com/rkb/bank/CheckingAccount.java
@@ -3,20 +3,20 @@ package com.rkb.bank;
 public class CheckingAccount extends BankingAccount {
 	private final int minimumBalance = 100;
 	
-	public CheckingAccount(long balance) {
+	public CheckingAccount(double balance) {
 		super();
 		this.setBalance(balance);
 		
 	}
 	
-	public void deopsit(long amount) {
-		long currentBalance = this.getBalance() + amount;
+	public void deopsit(double amount) {
+		double currentBalance = this.getBalance() + amount;
 		this.setBalance(currentBalance);
 	}
 	
-	public void withdraw(long amount) {
+	public void withdraw(double amount) {
 		if((this.getBalance() - this.minimumBalance) >= amount) {
-			long currentBalance = this.getBalance() + amount;
+			double currentBalance = this.getBalance() + amount;
 			this.setBalance(currentBalance);
 		}
 	}
