+++ src/com/rkb/bank/SavingAccount.java
@@ -3,19 +3,19 @@ package com.rkb.bank;
 public class SavingAccount extends BankingAccount{
 	private final int minimumBalance = 1000;
 	
-	public SavingAccount(long balance) {
+	public SavingAccount(double balance) {
 		super();
 		this.setBalance(balance);
 	}
 	
-	public void deopsit(long amount) {
-		long currentBalance = this.getBalance() + amount;
+	public void deopsit(double amount) {
+		double currentBalance = this.getBalance() + amount;
 		this.setBalance(currentBalance);
 	}
 	
-	public void withdraw(long amount) {
+	public void withdraw(double amount) {
 		if((this.getBalance() - this.minimumBalance) >= amount) {
-			long currentBalance = this.getBalance() + amount;
+			double currentBalance = this.getBalance() + amount;
 			this.setBalance(currentBalance);
 		}
 	}
